Plot heatmap key

heatmap.key plots a heatmap key in the current viewport.

heatmap.key(x, color.scheme = heatmap.color.scheme(),
  key.min = utils::head(color.scheme$breaks, 1) - 1,
  key.max = utils::tail(color.scheme$breaks, 1) + 1,
  key.line.color = "cyan")

Arguments

x
numeric matrix that was used to plot a heatmap

color.scheme
the color scheme used to plot the heatmap

key.min
should be less than the least break point of the color scheme

key.max
should be greater than the greates break point of the color scheme

key.line.color
the color used to draw the histogram

Value

invisible(NULL)