Patient linear ordering based on gene set expression

sig.ranksum linearly orders samples based on expression of a defined set of genes. Patients with the same ranksum are assigned their 'average' rank, i.e., they all receive the same rank value, which might not be an integer.

Genes in the set are partitioned into two groups around myeloids (ğ‘€1 and ğ‘€2) using correlation as the distance metric. Each gene in ğ‘€1 and ğ‘€2 is ordered from high to low and low to high expression, respectively. Expression values of each gene are then replaced by their ranks across patients. The sum of gene ranks (ranksum) are then used to linearly ordered patients. Note that 'up' and 'dn' or 'ns' are used to index into exprdata.

sig.ranksum(x.dat, up = NULL, dn = NULL, ns = NULL, n = 1000,
  middle.range = 0.95, seed = 123456, mc.cores = 2)

Arguments

x.dat
a list with at least 3 objects. exprs: gene expression data, must be a matrix. Genes are in rows, patients in columns. NAs in exprdata are not supported. clinical: clinical information data frame. cohorts: list of patient subgroups

up
indices of up-regulated genes

dn
indices of down-regulated genes

ns
indices of genes whose directions have not been specified.

n
number of permutations to perform to compute the region of indendence

middle.range
percentile point of the distribution of random patient ranks to define the region of independence

seed
random seed number

mc.cores
number of cores to use

Value

The output is a list containing the following objects

pat.order

vector providing the patient linear ordering. data[, pat.order] orders the columns of expression matrix according to the sum of gene ranks (gene ranksums).

ranksum

vector of gene ranksums acr

mgene.ordergene ordering. exprdata[gene.order, ] sorts rows of exprdata so that genes within each group around the two myeloids are ordered according to the strength of their association with the patient linear ordering. Most correlated genes are at each end of the ordering
dat

matrix of expression of up and dn or ns genes reordered by patient and gene ordering.

up.dn

vector with nrow(dat) values. -1, indicates down-gene and 1 indicates up-gene.

up

vector of indexes pointing at M1 genes in exprdata defined as upregulated i.e positively correlated with patient ordering.

dn

vector of indexes pointing at M2 genes in exprdata defined as downregulated i.e negatively correlated with patient ordering.